feat(hawtio-springboot): ENTESB-12346 - move reusable configs from examples into hawtio-springboot-keycloak
Fixes #2614
