+++ bom/pom.xml
@@ -82,6 +82,11 @@
       </dependency>
       <dependency>
         <groupId>io.hawt</groupId>
+        <artifactId>hawtio-springboot-keycloak</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.hawt</groupId>
         <artifactId>hawtio-system</artifactId>
         <version>${project.version}</version>
       </dependency>
