+++ examples/springboot-keycloak/src/main/resources/application.properties
@@ -9,11 +9,6 @@ logging.level.io.hawt=INFO
 logging.level.org.springframework=WARN
 logging.level.io.undertow=WARN
 
-# Hawtio Keycloak config
-hawtio.authenticationEnabled=true
-hawtio.keycloakEnabled=true
-hawtio.keycloakClientConfig=classpath:keycloak-hawtio.json
-
 # Keycloak config
 keycloak.realm=hawtio-demo
 keycloak.resource=hawtio-client
