+++ platforms/springboot-keycloak/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+io.hawt.springboot.keycloak.HawtioKeycloakConfiguration,\
+io.hawt.springboot.keycloak.KeycloakConfigResolverConfiguration
