+++ platforms/springboot-keycloak/src/main/resources/io/hawt/springboot/keycloak/application.properties
@@ -1,0 +1,3 @@
+hawtio.authenticationEnabled=true
+hawtio.keycloakEnabled=true
+hawtio.keycloakClientConfig=classpath:keycloak-hawtio.json
